requirements.txt
Django==4.0.1
django-environ==0.8.1
Pillow==9.0.0
requests==2.27.1
beautifulsoup4==4.10.0
